+++ flash/README.md
@@ -1,4 +1,4 @@
-# MVP
+# flash
 
 ## Examples
