+++ flash/prompts/core-concepts.txt
