+++ flash/prompts/real-life-examples.txt
