+++ flash/text-examples/docker/img/01-primary-blue-docker-logo.png
